Make all amounts in Account use BigDecimal

+++ app/src/main/java/com/concordium/wallet/data/model/AccountBalanceInfo.kt
@@ -1,6 +1,8 @@
 package com.concordium.wallet.data.model
 
+import com.concordium.wallet.util.toBigDecimal
 import java.io.Serializable
+import java.math.BigDecimal
 
 data class AccountBalanceInfo(
     val accountAmount: String,
@@ -12,8 +14,8 @@ data class AccountBalanceInfo(
     val accountIndex: Int
 ) : Serializable {
 
-    fun getAmount(): Long {
-        return accountAmount.toLong()
+    fun getAmount(): BigDecimal {
+        return accountAmount.toBigDecimal()
     }
 
     fun getEncryptedAmount(): AccountEncryptedAmount {
