+++ app/src/main/java/com/concordium/wallet/ui/account/common/accountupdater/TotalBalancesData.kt
@@ -1,8 +1,10 @@
 package com.concordium.wallet.ui.account.common.accountupdater
 
+import java.math.BigDecimal
+
 data class TotalBalancesData(
-    val totalBalanceForAllAccounts: Long,
-    val totalAtDisposalForAllAccounts: Long,
-    val totalStakedForAllAccounts: Long,
+    val totalBalanceForAllAccounts: BigDecimal,
+    val totalAtDisposalForAllAccounts: BigDecimal,
+    val totalStakedForAllAccounts: BigDecimal,
     val totalContainsEncrypted: Boolean
 )
